+++ src/pytorch_lightning/strategies/deepspeed.py
@@ -89,7 +89,7 @@ class LightningDeepSpeedModule(_LightningModuleWrapperBase):
 
     def _batch_to(self, batch: Tensor) -> Tensor:
         if torch.is_floating_point(batch):
-            if self.precision in (PrecisionType.HALF, PrecisionType.MIXED):
+            if self.precision == PrecisionType.HALF:
                 return batch.half()
             elif self.precision == PrecisionType.BFLOAT:
                 return batch.bfloat16()
@@ -547,7 +547,7 @@ class DeepSpeedStrategy(DDPStrategy):
         if self.zero_stage_3:
             assert self._config_initialized
 
-            if self.precision_plugin.precision in (PrecisionType.HALF, PrecisionType.MIXED):
+            if self.precision_plugin.precision == PrecisionType.HALF:
                 dtype = torch.float16
             elif self.precision_plugin.precision == PrecisionType.BFLOAT:
                 dtype = torch.bfloat16
@@ -696,7 +696,7 @@ class DeepSpeedStrategy(DDPStrategy):
 
     def _format_precision_config(self) -> None:
         assert isinstance(self.config, dict)
-        if self.precision_plugin.precision in (PrecisionType.HALF, PrecisionType.MIXED):
+        if self.precision_plugin.precision == PrecisionType.HALF:
             if "fp16" not in self.config and self.precision_plugin.amp_type == AMPType.NATIVE:
                 # FP16 is a DeepSpeed standalone AMP implementation
                 rank_zero_info("Enabling DeepSpeed FP16.")